requirements.txt
argon2-cffi==19.1.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.0.2
django-extensions==2.2.9
django-guardian==2.0.0
django-postgres-extra==1.22
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==2.0.1
drf-yasg==1.16.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.1
psycopg2==2.8.3
pycparser==2.19
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
